+++ src/app/src/components/FacilityListItemsTable.jsx
@@ -51,6 +51,7 @@ import {
     facilityListItemStatusChoicesEnum,
     facilityListItemErrorStatuses,
     facilityListStatusFilterChoices,
+    matchResponsibilityEnum,
 } from '../util/constants';
 
 const facilityListItemsTableStyles = Object.freeze({
@@ -700,7 +701,6 @@ class FacilityListItemsTable extends Component {
 
 FacilityListItemsTable.defaultProps = {
     items: null,
-    readOnly: false,
     errorRemovingItem: null,
 };
 
@@ -720,7 +720,7 @@ FacilityListItemsTable.propTypes = {
     makeRemoveFacilityListItemFunction: func.isRequired,
     isRemovingItem: bool.isRequired,
     errorRemovingItem: arrayOf(string),
-    readOnly: bool,
+    readOnly: bool.isRequired,
 };
 
 function mapStateToProps({
@@ -738,6 +738,18 @@ function mapStateToProps({
         user: { user },
     },
 }) {
+    const isAdminUser =
+        !!user &&
+        user.is_superuser &&
+        !!list &&
+        user.contributor_id !== list.contributor_id;
+    const readOnly =
+        (!!list &&
+            list.match_responsibility === matchResponsibilityEnum.CONTRIBUTOR &&
+            isAdminUser) ||
+        (!!list &&
+            list.match_responsibility === matchResponsibilityEnum.MODERATOR &&
+            !isAdminUser);
     return {
         list,
         items,
@@ -746,11 +758,7 @@ function mapStateToProps({
         isRemovingItem,
         errorRemovingItem,
         selectedFacilityListItemsRowIndex,
-        readOnly:
-            user &&
-            user.is_superuser &&
-            list &&
-            user.contributor_id !== list.contributor_id,
+        readOnly,
     };
 }