+++ src/app/src/util/constants.js
@@ -54,6 +54,17 @@ export const facilityClaimStatusChoicesEnum = Object.freeze({
 });
 
 // These choices must be kept in sync with the identical list
+// kept in the Django API's constants.py file
+export const matchResponsibilityEnum = Object.freeze({
+    MODERATOR: 'moderator',
+    CONTRIBUTOR: 'contributor',
+});
+export const matchResponsibilityChoices = [
+    { value: 'moderator', label: 'OS Hub Admins' },
+    { value: 'contributor', label: 'The contributor' },
+];
+
+// These choices must be kept in sync with the identical list
 // kept in the Django API's models.py file
 export const contributorWebhookNotificationChoices = [
     { value: 'ALL_FACILITIES', label: 'All events' },