+++ src/django/api/constants.py
@@ -47,6 +47,8 @@ class FacilitiesQueryParams:
 
 class FacilityListQueryParams:
     CONTRIBUTOR = 'contributor'
+    STATUS = 'status'
+    MATCH_RESPONSIBILITY = 'match_responsibility'
 
 
 class FacilityListItemsQueryParams:
@@ -70,6 +72,10 @@ class FeatureGroups:
     CAN_SUBMIT_FACILITY = 'can_submit_facility'
     CAN_SUBMIT_PRIVATE_FACILITY = 'can_submit_private_facility'
     CAN_VIEW_FULL_CONTRIB_DETAIL = 'can_view_full_contrib_detail'
+
+
+class FacilityListStatus:
+    MATCHED = "MATCHED"
 
 
 class FacilityHistoryActions:
@@ -155,6 +161,16 @@ class DateFormats:
     SECOND = '%Y-%m-%d %H:%M:%S'
     MONTH = '%Y-%m'
     WEEK = '%Y-%W'
+
+
+class MatchResponsibility:
+    MODERATOR = "moderator"
+    CONTRIBUTOR = "contributor"
+
+    CHOICES = [
+        (MODERATOR, 'OS Hub admins'),
+        (CONTRIBUTOR, 'The contributor'),
+    ]
 
 
 class NumberOfWorkersRanges: