+++ src/django/api/tests.py
@@ -24,7 +24,9 @@ from rest_framework.authtoken.models import Token
 from rest_framework.test import APITestCase
 from waffle.testutils import override_switch, override_flag
 
-from api.constants import (FacilityHistoryActions, ProcessingAction,
+from api.constants import (FacilityHistoryActions,
+                           ProcessingAction,
+                           MatchResponsibility,
                            LogDownloadQueryParams,
                            UpdateLocationParams,
                            FeatureGroups)
@@ -1475,7 +1477,8 @@ class ConfirmRejectAndRemoveAndDissociateFacilityMatchTest(TestCase):
             .objects \
             .create(header="header",
                     file_name="one",
-                    name=self.current_list_name)
+                    name=self.current_list_name,
+                    match_responsibility=MatchResponsibility.CONTRIBUTOR)
 
         self.current_source = Source \
             .objects \
@@ -2962,7 +2965,7 @@ class FacilityClaimTest(APITestCase):
         self.assertEqual(1, len(data))
 
 
-class DashboardListTests(APITestCase):
+class BaseFacilityListTests(APITestCase):
     def setUp(self):
         self.user_email = 'test@example.com'
         self.user_password = 'example123'
@@ -3055,22 +3058,51 @@ class DashboardListTests(APITestCase):
             facility_list=self.superlist,
             contributor=self.supercontributor)
 
-    def test_user_can_list_own_lists(self):
-        self.client.login(email=self.user_email,
-                          password=self.user_password)
-        response = self.client.get('/api/facility-lists/')
+
+class DashboardListTests(BaseFacilityListTests):
+    def test_superuser_can_list_all_lists(self):
+        self.client.login(email=self.superuser_email,
+                          password=self.superuser_password)
+        response = self.client.get('/api/admin-facility-lists/')
 
         self.assertEqual(200, response.status_code)
 
         lists = response.json()
 
-        # Ensure we get all three lists
-        self.assertEqual(3, len(lists))
-        # Ensure they are ordered newest first
-        self.assertEqual(
-            ['Third List', 'Second List', 'First List'],
-            [d['name'] for d in lists])
+        # Ensure we get the all lists
+        self.assertEqual(4, len(lists['results']))
 
+    def test_superuser_can_list_other_contributors_lists(self):
+        self.client.login(email=self.superuser_email,
+                          password=self.superuser_password)
+        response = self.client.get(
+            '/api/admin-facility-lists/?contributor={}'.format(
+                self.contributor.id))
+
+        self.assertEqual(200, response.status_code)
+
+        lists = response.json()
+
+        # Ensure we get all three lists, private and public,
+        # active and inactive
+        self.assertEqual(3, len(lists['results']))
+        self.assertEqual(
+            ['First List', 'Second List', 'Third List'],
+            [d['name'] for d in lists['results']])
+
+    def test_user_cannot_list_other_contributors_lists(self):
+        # Regular users cannot access this endpoint
+
+        self.client.login(email=self.user_email,
+                          password=self.user_password)
+        response = self.client.get(
+            '/api/admin-facility-lists/?contributor={}'.format(
+                self.supercontributor.id))
+
+        self.assertEqual(403, response.status_code)
+
+
+class FacilityListViewTests(BaseFacilityListTests):
     def test_superuser_can_list_own_lists(self):
         self.client.login(email=self.superuser_email,
                           password=self.superuser_password)
@@ -3084,44 +3116,6 @@ class DashboardListTests(APITestCase):
         self.assertEqual(1, len(lists))
         # Ensure it is the right one
         self.assertEqual('Super List', lists[0]['name'])
-
-    def test_superuser_can_list_other_contributors_lists(self):
-        self.client.login(email=self.superuser_email,
-                          password=self.superuser_password)
-        response = self.client.get(
-            '/api/facility-lists/?contributor={}'.format(
-                self.contributor.id))
-
-        self.assertEqual(200, response.status_code)
-
-        lists = response.json()
-
-        # Ensure we get all three lists, private and public,
-        # active and inactive
-        self.assertEqual(3, len(lists))
-        self.assertEqual(
-            ['Third List', 'Second List', 'First List'],
-            [d['name'] for d in lists])
-
-    def test_user_cannot_list_other_contributors_lists(self):
-        # Regular users, even if they ask for lists by other
-        # contributors, will just be given their own lists
-
-        self.client.login(email=self.user_email,
-                          password=self.user_password)
-        response = self.client.get(
-            '/api/facility-lists/?contributor={}'.format(
-                self.supercontributor.id))
-
-        self.assertEqual(200, response.status_code)
-
-        lists = response.json()
-
-        # Ensure we get the user's, not the superuser's, lists
-        self.assertEqual(3, len(lists))
-        self.assertEqual(
-            ['Third List', 'Second List', 'First List'],
-            [d['name'] for d in lists])
 
     def test_user_can_view_own_lists(self):
         self.client.login(email=self.user_email,
@@ -5393,7 +5387,8 @@ class FacilityHistoryEndpointTest(FacilityAPITestCaseBase):
             .objects \
             .create(header='header',
                     file_name='two',
-                    name='Second List')
+                    name='Second List',
+                    match_responsibility=MatchResponsibility.CONTRIBUTOR)
 
         self.source_two = Source \
             .objects \
@@ -5435,7 +5430,8 @@ class FacilityHistoryEndpointTest(FacilityAPITestCaseBase):
             .objects \
             .create(header='List for confirm or reject',
                     file_name='list for confirm or reject',
-                    name='List for confirm or reject')
+                    name='List for confirm or reject',
+                    match_responsibility=MatchResponsibility.CONTRIBUTOR)
 
         self.source_for_confirm_or_remove = Source \
             .objects \
@@ -7039,10 +7035,10 @@ class SingleItemFacilityMatchTest(FacilityAPITestCaseBase):
         self.assertEqual(200, response.status_code)
 
     def test_only_contributor_can_get_match_detail(self):
-        self.client.login(email=self.superuser_email,
-                          password=self.superuser_password)
+        self.client.login(email=self.user_email,
+                          password=self.user_password)
 
-        response = self.client.get(self.match_url(self.match))
+        response = self.client.get(self.match_url(self.match_two))
         self.assertEqual(404, response.status_code)
 
     def test_confirm(self):
@@ -7054,11 +7050,11 @@ class SingleItemFacilityMatchTest(FacilityAPITestCaseBase):
         self.assertEqual(200, response.status_code)
 
     def test_only_contributor_can_confirm(self):
-        self.client.login(email=self.superuser_email,
-                          password=self.superuser_password)
+        self.client.login(email=self.user_email,
+                          password=self.user_password)
 
         response = self.client.post(
-            self.match_url(self.match, action='confirm'))
+            self.match_url(self.match_two, action='confirm'))
         self.assertEqual(404, response.status_code)
 
     def test_reject(self):
@@ -7069,10 +7065,10 @@ class SingleItemFacilityMatchTest(FacilityAPITestCaseBase):
         self.assertEqual(200, response.status_code)
 
     def test_only_contributor_can_reject(self):
-        self.client.login(email=self.superuser_email,
-                          password=self.superuser_password)
+        self.client.login(email=self.user_email,
+                          password=self.user_password)
         response = self.client.post(
-            self.match_url(self.match, action='reject'))
+            self.match_url(self.match_two, action='reject'))
         self.assertEqual(404, response.status_code)
 
 
@@ -7304,7 +7300,8 @@ class PPEFieldTest(TestCase):
             .objects \
             .create(header='header',
                     file_name='two',
-                    name='list_two')
+                    name='list_two',
+                    match_responsibility=MatchResponsibility.CONTRIBUTOR)
 
         self.source_one = Source \
             .objects \